requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
dgl
torchmetrics
sacred
tensorboard
jarvis-tools
pymatgen
pandas
jupyterlab
pylint
matbench
fire
botocore
transformers
# torch and pytorch-lightning
# torch==2.0.1
# pytorch-lightning==2.0.6
# dgl 
# pip install  dgl -f https://data.dgl.ai/wheels/cu117/repo.html
# pip install  dglgo -f https://data.dgl.ai/wheels-test/repo.html